Adding fingerprint data for case study, UI changes, and Bug Fixes
A data related to a fingerprint is uploaded as .png, .jpg, .jpeg, or pdf formats. Uploaded files can only be deleted after uploading. No updates can be performed.

The UI was changed as a side by side tabbed format to utilize available space efficiently.

Error msgs related to file uploading  were made more informative.
Evidence table returns null records when searching fields due to as all data are not available per row. "IS NOT NULL" were used for queries associated with the table to avoid this.